[stereo-mix] Add a stereo mix server
A stereo mix server is a local TCP server that outputs the mixed sound as s16le stereo @ 48000 hz.

The output can be streamed to ffmpeg for sending to (e.g.) an Icecast server:

     nc localhost $STEREO_MIX_PORT \
       | ffmpeg -f s16le -ar 48000 -ac 2 -i - \
                -acodec libmp3lame -ab 128k -f mp3 $ICECAST_URL

The implementation for `CServer::MixStream` comes from the Streamer2 PR:
https://github.com/jamulussoftware/jamulus/pull/967

By making the stream system pull-based (consumers ask Jamulus for data),
rather than push-based (Jamulus sends data to a streaming server), the
implementation is greatly simplified.

* No need to implement ways to toggle streaming on/off, as streaming
  starts when a client connects and ends when the client disconnects.
* No need to deal with launching, monitoring, and killing sub-processes.
* Works with all OSes.

Co-authored-by: dingodoppelt <dexxter@top-email.net>